+++ moonraker/components/update_manager/app_deploy.py
@@ -11,11 +11,11 @@ import shutil
 import hashlib
 import logging
 import re
-import json
 import distro
 import asyncio
 from .common import AppType, Channel
 from .base_deploy import BaseDeploy
+from ...utils import json_wrapper as jsonw
 
 # Annotation imports
 from typing import (
@@ -278,7 +278,7 @@ class AppDeploy(BaseDeploy):
             deps_json = self.system_deps_json
             try:
                 ret = await eventloop.run_in_thread(deps_json.read_bytes)
-                dep_info: Dict[str, List[str]] = json.loads(ret)
+                dep_info: Dict[str, List[str]] = jsonw.loads(ret)
             except asyncio.CancelledError:
                 raise
             except Exception:
@@ -482,7 +482,7 @@ class AppDeploy(BaseDeploy):
             await self.cmd_helper.run_cmd(
                 f"virtualenv {self.venv_args} {env_path}", timeout=300.)
         except Exception:
-            self.log_exc(f"Error creating virtualenv")
+            self.log_exc("Error creating virtualenv")
             return
         if not self.py_exec.exists():
             raise self.log_exc("Failed to create new virtualenv", False)