+++ scripts/install-moonraker.sh
@@ -12,8 +12,8 @@ CONFIG_PATH="${MOONRAKER_CONFIG_PATH}"
 LOG_PATH="${MOONRAKER_LOG_PATH}"
 DATA_PATH="${MOONRAKER_DATA_PATH}"
 INSTANCE_ALIAS="${MOONRAKER_ALIAS:-moonraker}"
+SPEEDUPS="${MOONRAKER_SPEEDUPS:-n}"
 SERVICE_VERSION="1"
-MACHINE_PROVIDER="systemd_cli"
 
 package_decode_script=$( cat << EOF
 import sys
@@ -88,6 +88,11 @@ create_virtualenv()
 
     # Install/update dependencies
     ${PYTHONDIR}/bin/pip install -r ${SRCDIR}/scripts/moonraker-requirements.txt
+
+    if [ ${SPEEDUPS} = "y" ]; then
+        report_status "Installing Speedups..."
+        ${PYTHONDIR}/bin/pip install -r ${SRCDIR}/scripts/moonraker-speedups.txt
+    fi
 }
 
 # Step 5: Initialize data folder
@@ -105,6 +110,12 @@ init_data_path()
     [ -n "${CONFIG_PATH}" ] && config_file=${CONFIG_PATH}
     # Write initial configuration for first time installs
     if [ ! -f $SERVICE_FILE ] && [ ! -e "${config_file}" ]; then
+        # detect machine provider
+        if [ "$( systemctl is-active dbus )" = "active" ]; then
+            provider="systemd_dbus"
+        else
+            provider="systemd_cli"
+        fi
         report_status "Writing Config File ${config_file}:\n"
         /bin/sh -c "cat > ${config_file}" << EOF
 # Moonraker Configuration File
@@ -117,7 +128,7 @@ port: 7125
 klippy_uds_address: /tmp/klippy_uds
 
 [machine]
-provider: ${MACHINE_PROVIDER}
+provider: ${provider}
 
 EOF
         cat ${config_file}
@@ -171,7 +182,7 @@ EOF
 # Step 7: Validate/Install polkit rules
 check_polkit_rules()
 {
-    if [ ! -x "$(command -v pkaction)" ]; then
+    if [ ! -x "$(command -v pkaction || true)" ]; then
         return
     fi
     POLKIT_VERSION="$( pkaction --version | grep -Po "(\d+\.?\d*)" )"
@@ -198,10 +209,7 @@ check_polkit_rules()
         else
             report_status "Installing PolKit Rules"
             ${SRCDIR}/scripts/set-policykit-rules.sh -z
-            MACHINE_PROVIDER="systemd_dbus"
         fi
-    else
-        MACHINE_PROVIDER="systemd_dbus"
     fi
 }
 
@@ -233,12 +241,13 @@ set -e
 SRCDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )"/.. && pwd )"
 
 # Parse command line arguments
-while getopts "rfzxc:l:d:a:" arg; do
+while getopts "rfzxsc:l:d:a:" arg; do
     case $arg in
         r) REBUILD_ENV="y";;
         f) FORCE_DEFAULTS="y";;
         z) DISABLE_SYSTEMCTL="y";;
         x) SKIP_POLKIT="y";;
+        s) SPEEDUPS="y";;
         c) CONFIG_PATH=$OPTARG;;
         l) LOG_PATH=$OPTARG;;
         d) DATA_PATH=$OPTARG;;